Fixing painted white conversion

+++ composeApp/src/commonMain/kotlin/dev/anthonyhfm/amethyst/conversion/ableton/data/LiveSetData.kt
@@ -254,7 +254,7 @@ data class MidiClip(
                         val duration: Float,
 
                         @XmlSerialName("Velocity")
-                        val velocity: Int,
+                        val velocity: Float,
                     )
                 }
             }
